requirements.txt
sentence-transformers
psycopg2
numpy
openai
llama-cpp-python
python-dotenv